Move screen name mapping to a separate class
